Move the code in app/src/context/theme.ts & updated dependencies.
