+++ app/src/App.tsx
@@ -1,5 +1,5 @@
 import React, { useCallback, useEffect, useState } from 'react';
-import {Theme,ThemeContext} from './theme/themeContext'
+import {Theme,ThemeContext} from './context/theme'
 import ActionToolbar from './features/toolbar/components/ActionToolbar';
 import LogView from './features/editor/components/LogView';
 import { useCompile } from './features/editor/hooks/useCompile';
