+++ app/src/components/Copyable.tsx
@@ -3,7 +3,7 @@ import { darkColors } from '@fuel-ui/css';
 import IconButton from '@mui/material/IconButton';
 import Tooltip from '@mui/material/Tooltip';
 import ContentCopyIcon from '@mui/icons-material/ContentCopy';
-import { useThemeContext } from "../theme/themeContext";
+import { useThemeContext } from "../context/theme";
 
 export interface CopyableProps {
   value: string;
